requirements.txt
fastapi
uvicorn
redis
faiss-cpu
sentence-transformers
rank-bm25
langchain
PyPDF2